Fix PD test to work on slower/busier systems.
The assert_needs(self, need_counts, domain_id=None) function in mocks
assumes that for each process submitted, there will be one matchmaking
cycle. This will not be true on slower/busier systems. The test has been
changed to only check the last EPUM reconfigure, ensuring that it requests
the correct number of nodes for the number of processes submitted