Retire old node records in terminal state in provisioner
Node recors in states TERMINATED, FAILED, or REJECTED are deleted if
they are older than record_reaping_max_age.  The default configuration
is 2 hours and can be changed through the provisioner YAML configuration
file (it needs to be expressed in seconds).

We also delete launches when they do not have any more node records
after removal.