Delete engines only if they already exist
Since engines are created asynchronously, it is possible for a domain to
be shut down before its engine has been created. This would trigger an
exception that would kill the main decider loop, which would be
restarted only 5 seconds later. This problem happened repetitively in
TestIntegrationDomain.many_domain_simple_test.