Update node before resource on termination in heartbeat.
This avoids the following race:

with node_exclusive, when a process terminates it triggers the MM loop
before the node_exclusive flag is cleared on the node

so like if you have processes A and B with node_exclusive="foo". A is
running and B is in the queue. if A dies, the update of the resource
record will trigger the MM loop. But that can happen before the
node_exclusive flag is cleared off the node.