Refresh node IP information even when the instance state stays the same
This is a workaround for OpenStack, which returns the local hostname as
the public hostname for some time during the launch, even though the
instance is already running. At each IaaS query, we check if the
instance IP information has changed. If it does, we send a state update
to EPUM.

We make sure that EPUM does not process out of order messages by using
an update counter.